requirements.txt
fastapi
uvicorn[standard]
fpdf2
streamlit
pandas
matplotlib
numpy
scipy
plotly
openpyxl
kaleido
